Move share planning behind wallet boundary
