+++ rust/src/wallet/voting.rs
@@ -2,6 +2,7 @@ pub mod db;
 pub mod delegation;
 pub mod hotkey;
 pub mod network;
+pub mod share;
 
 #[cfg(test)]
 pub(crate) mod test_support;
